requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
tensorflow==2.13.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
transformers==4.33.2
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
# Additional dependencies for better compatibility
Keras==2.13.1
h5py==3.8.0
joblib==1.3.2
